move mobilenav to ui/mobile
